refactor: move PageBase from layout to ui
