+++ src/components/layouts/index.ts
@@ -1,4 +1,3 @@
-import PageBase from "./PageBase";
 import PageHead from "./PageHead";
 import type { PageHeadProps } from "./PageHead";
 import PolicyPageLayout from "./PolicyPageLayout";
@@ -6,5 +5,5 @@ import type { PolicyPageLayoutProps } from "./PolicyPageLayout";
 import StickyScrollLayout from "./StickyScrollLayout";
 import type { StickyScrollLayoutProps } from "./StickyScrollLayout";
 
-export { PageBase, PageHead, PolicyPageLayout, StickyScrollLayout };
+export { PageHead, PolicyPageLayout, StickyScrollLayout };
 export type { PageHeadProps, PolicyPageLayoutProps, StickyScrollLayoutProps };
