+++ src/components/ui/PageBase.tsx
@@ -1,12 +1,16 @@
 import dynamic from "next/dynamic";
-import { FC } from "react";
+import { FC, ReactNode } from "react";
 import { Nav } from "@/components/ui";
 
 const Footer = dynamic(() =>
   import("@/components/ui").then((mod) => mod.Footer)
 );
 
-const PageBase: FC = ({ children }) => {
+export type PageBaseProps = {
+  children?: ReactNode | undefined;
+};
+
+const PageBase: FC = ({ children }: PageBaseProps) => {
   return (
     <div className="flex min-h-screen flex-col bg-instillGrey95">
       <Nav />
