+++ src/pages/get-access.tsx
@@ -1,8 +1,9 @@
 import { FC, ReactElement } from "react";
 import Image from "next/future/image";
 
-import { PageBase, PageHead } from "@/components/layouts";
+import { PageHead } from "@/components/layouts";
 import { MailchimpSignupForm } from "../components/forms/MailchimpSignupForm";
+import { PageBase } from "@/components/ui";
 
 interface GetLayOutProps {
   page: ReactElement;
